+++ Samples/Cordova/ExampleProject/platforms/android/platform_www/cordova_plugins.js
@@ -1,24 +1,6 @@
 cordova.define('cordova/plugin_list', function(require, exports, module) {
-  module.exports = [
-    {
-      "id": "cordova-plugin-x-toast.Toast",
-      "file": "plugins/cordova-plugin-x-toast/www/Toast.js",
-      "pluginId": "cordova-plugin-x-toast",
-      "clobbers": [
-        "window.plugins.toast"
-      ]
-    },
-    {
-      "id": "clevertap-cordova.CleverTap",
-      "file": "plugins/clevertap-cordova/www/CleverTap.js",
-      "pluginId": "clevertap-cordova",
-      "clobbers": [
-        "CleverTap"
-      ]
-    }
-  ];
+  module.exports = [];
   module.exports.metadata = {
-    "cordova-plugin-x-toast": "2.7.3",
-    "clevertap-cordova": "3.4.0"
+    "cordova-android-support-gradle-release": "3.0.1"
   };
 });