+++ Samples/IonicCordova/IonicAngularProject/src/app/home/home.page.ts
@@ -435,6 +435,13 @@ export class HomePage {
         this.clevertap.pushInboxNotificationClickedEventForId('Insert message id');
     }
 
+    fetchInbox() {
+        console.log('fetchInbox');
+        (this.clevertap as any).fetchInbox().then((r: boolean) => {
+            this.presentToast('fetchInbox: ' + r);
+        });
+    }
+
     suspendInAppNotifications() {
         console.log('suspendInAppNotifications');
         this.clevertap.suspendInAppNotifications();
@@ -472,6 +479,12 @@ export class HomePage {
         console.log('pushDisplayUnitClickedEventForID');
         this.clevertap.pushDisplayUnitClickedEventForID('Test Display Unit');
         this.presentToast('pushDisplayUnitClickedEventForID');
+    }
+
+    pushDisplayUnitElementClickedEventForID() {
+        console.log('pushDisplayUnitElementClickedEventForID');
+        (this.clevertap as any).pushDisplayUnitElementClickedEventForID('Test Display Unit', { wzrk_element_id: 'btn_01' });
+        this.presentToast('pushDisplayUnitElementClickedEventForID');
     }
 
     getFeatureFlag() {